refactor: sharedBaseApi hander
