+++ packages/nocodb/src/lib/services/index.ts
@@ -19,3 +19,4 @@ export * as viewColumnService from './viewColumnService';
 export * as metaDiffService from './metaDiffService';
 export * as mapViewService from './mapViewService';
 export * as modelVisibilityService from './modelVisibilityService';
+export * as sharedBaseService from './sharedBaseService';
